Tests/LibWeb/Text/expected/Worker/Worker-requestAnimationFrame.txt

has requestAnimationFrame: true
has cancelAnimationFrame: true
callback ran with numeric timestamp: true
second frame not before first: true
